pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
ta = "^0.11.0"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
kagglehub = "^0.2.9"
psycopg2-binary = "2.9.9"
huggingface-hub = "0.24.5"
transformers = "4.43.3"
datasets = "2.20.0"
gradio = "4.40.0"
gradio-agentchatbot = "0.0.8"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"